Proposal: break gomod code into separate modules
Gomod code has grown way to big. With the recent changes to fully
generate the SBOM, I can see three distinct responsibilites:

- parser: calls and handles output from go commands and go files
- resolver: turns the output from gomod.parser into a valid sbom
- version: handles versions defined in the Git repository
